test: add local browser tests with playwright
